cherry pick #11086 to release-4.0
close #11082
