+++ components/cdc/src/lib.rs
@@ -22,9 +22,9 @@ mod metrics;
 mod observer;
 mod service;
 
-pub use channel::{recv_timeout, MemoryQuota};
+pub use channel::{recv_timeout, CdcEvent, MemoryQuota};
 pub use config::CdcConfigManager;
 pub use endpoint::{Endpoint, OldValueStats, Task, Validate};
 pub use errors::{Error, Result};
 pub use observer::CdcObserver;
-pub use service::{CdcEvent, FeatureGate, Service};
+pub use service::{FeatureGate, Service};
